History in DB switch

+++ internal/web/trim-history.go
@@ -5,6 +5,7 @@ import (
 	"time"
 
 	"github.com/aceberg/WatchYourLAN/internal/db"
+	"github.com/aceberg/WatchYourLAN/internal/models"
 )
 
 func trimHistoryRoutine() {
@@ -24,19 +25,35 @@ func trimHistory() {
 	now, _ := time.Parse("2006-01-02 15:04:05", nowStr) // in one format
 	nowMinus := now.Add(-time.Duration(hours) * time.Hour)
 
-	history := db.Select("history")
+	history := histHosts
+	if appConfig.HistInDB {
+		history = db.Select("history")
+	}
 
 	slog.Info("Removing all History before", "date", nowMinus.Format("2006-01-02 15:04:05"))
 
 	n := 0
-	for _, hist := range history {
+	for i, hist := range history {
 		date, _ := time.Parse("2006-01-02 15:04:05", hist.Date)
 
 		if date.Before(nowMinus) {
 			n = n + 1
-			db.Delete("history", hist.ID)
+			if appConfig.HistInDB {
+				db.Delete("history", hist.ID)
+			} else {
+				historyDel(i)
+			}
 		}
 	}
 
 	slog.Info("Removed records from History", "n", n)
+}
+
+func historyDel(i int) {
+
+	l := len(histHosts)
+
+	histHosts[i] = histHosts[l-1]
+	histHosts[l-1] = models.Host{}
+	histHosts = histHosts[:l-1]
 }
